untangled package context from rook package (#409)
* untangled package context from rook package
