+++ commands/package_init.go
@@ -7,6 +7,7 @@ import (
 	"gopkg.in/urfave/cli.v1"
 
 	"github.com/Southclaws/sampctl/download"
+	"github.com/Southclaws/sampctl/pkgcontext"
 	"github.com/Southclaws/sampctl/print"
 	"github.com/Southclaws/sampctl/rook"
 	"github.com/Southclaws/sampctl/util"
@@ -33,7 +34,7 @@ func packageInit(c *cli.Context) error {
 		return err
 	}
 
-	_, err = rook.NewPackageContext(gh, gitAuth, true, dir, platform(c), cacheDir, "")
+	_, err = pkgcontext.NewPackageContext(gh, gitAuth, true, dir, platform(c), cacheDir, "")
 	if err == nil {
 		return errors.New("Directory already appears to be a package")
 	}
