+++ pawnpackage/tests/.gitignore
@@ -1,0 +1,5 @@
+cache/
+deps/
+deps-*
+*.amx
+build-auto-*
