+++ pkgcontext/build.go
@@ -1,4 +1,4 @@
-package rook
+package pkgcontext
 
 import (
 	"context"
@@ -17,7 +17,6 @@ import (
 
 	"github.com/fatih/color"
 	"github.com/fsnotify/fsnotify"
-	"github.com/imdario/mergo"
 	"github.com/pkg/errors"
 
 	"github.com/Southclaws/sampctl/build"
@@ -265,7 +264,7 @@ func (pcx *PackageContext) buildPrepare(
 	ensure,
 	forceUpdate bool,
 ) (config *build.Config, err error) {
-	config = GetBuildConfig(pcx.Package, build)
+	config = pcx.Package.GetBuildConfig(build)
 	if config == nil {
 		err = errors.Errorf("no build config named '%s'", build)
 		return
@@ -327,68 +326,6 @@ func (pcx *PackageContext) buildPrepare(
 	config.Includes = append(config.Includes, pcx.AllIncludePaths...)
 
 	return config, err
-}
-
-// GetBuildConfig returns a matching build by name from the package build list. If no name is
-// specified, the first build is returned. If the package has no build definitions, a default
-// configuration is returned.
-func GetBuildConfig(pkg pawnpackage.Package, name string) (config *build.Config) {
-	def := build.Default()
-
-	// if there are no builds at all, use default
-	if len(pkg.Builds) == 0 && pkg.Build == nil {
-		return def
-	}
-
-	// if the user did not specify a specific build config, use the first
-	// otherwise, search for a matching config by name
-	if name == "" {
-		if pkg.Build != nil {
-			config = pkg.Build
-		} else {
-			config = pkg.Builds[0]
-
-			if pkg.Build != nil {
-				mergo.Merge(&config, pkg.Builds[0])
-			}
-		}
-	} else {
-		for _, cfg := range pkg.Builds {
-			if cfg.Name == name {
-				config = cfg
-
-				if pkg.Build != nil {
-					mergo.Merge(config, pkg.Build)
-				}
-
-				break
-			}
-		}
-	}
-
-	if config == nil {
-		if pkg.Build != nil {
-			print.Warn("Build doesn't exist, defaulting to main build")
-			config = pkg.Build
-		} else {
-			print.Warn("No build config called:", name, "using default")
-			config = def
-		}
-	}
-
-	if config.Version != "" {
-		config.Compiler.Version = string(config.Version)
-	}
-
-	if config.Compiler.Version == "" {
-		config.Compiler.Version = def.Compiler.Version
-	}
-
-	if len(config.Args) == 0 {
-		config.Args = def.Args
-	}
-
-	return config
 }
 
 func readInt(file string) (n uint32, err error) {
