+++ pkgcontext/pkgcontext_test.go
@@ -1,4 +1,4 @@
-package rook
+package pkgcontext
 
 import (
 	"path/filepath"
