+++ pkgcontext/uninstall.go
@@ -1,4 +1,4 @@
-package rook
+package pkgcontext
 
 import (
 	"github.com/pkg/errors"
