--- rook/tests/load-yaml/pawn.yaml
@@ -1,5 +1,0 @@
-entry: gamemodes/test.pwn
-output: gamemodes/test.amx
-dependencies:
-  - sampctl/samp-stdlib:0.3.7-R2-2-1
-  - Southclaws/pawn-errors:1.2.3
